requirements.txt
amqp==2.5.1
attrs==19.2.0
Automat==0.7.0
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
constantly==15.1.0
Django==1.11.11
django-redis==4.10.0
dnspython==1.16.0
eventlet==0.25.1
fdfs-client-py==1.2.6
greenlet==0.4.15
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.23
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.10.1
kombu==4.6.5
MarkupSafe==1.1.1
monotonic==1.5
more-itertools==7.2.0
Pillow==6.2.0
PyHamcrest==1.9.0
PyMySQL==0.9.3
pytz==2019.2
QQLoginTool==0.3.0
redis==3.3.8
requests==2.22.0
six==1.12.0
Twisted==19.7.0
urllib3==1.25.6
vine==1.3.0
zipp==0.6.0
zope.interface==4.6.0
